Major update including support for Electrum wallet / Master Public Key functionality

+++ bitcoinway-woocommerce.php
@@ -11,7 +11,7 @@
 Plugin Name: Bitcoin Payments for WooCommerce
 Plugin URI: http://www.bitcoinway.com/
 Description: Bitcoin Payments for WooCommerce plugin allows you to accept payments in bitcoins for physical and digital products at your WooCommerce-powered online store.
-Version: 1.27
+Version: 2.02
 Author: BitcoinWay
 Author URI: http://www.bitcoinway.com/
 License: GNU General Public License 2.0 (GPL) http://www.gnu.org/licenses/gpl.html
@@ -33,6 +33,9 @@ register_activation_hook(__FILE__,          'BWWC_activate');
 register_deactivation_hook(__FILE__,        'BWWC_deactivate');
 register_uninstall_hook(__FILE__,           'BWWC_uninstall');
 
+add_filter ('cron_schedules',               'BWWC__add_custom_scheduled_intervals');
+add_action ('BWWC_cron_action',             'BWWC_cron_job_worker');     // Multiple functions can be attached to 'BWWC_cron_action' action
+
 BWWC_set_lang_file();
 //---------------------------------------------------------------------------
 
@@ -40,26 +43,48 @@ BWWC_set_lang_file();
 // activating the default values
 function BWWC_activate()
 {
-    global  $g_BWWC__config_defaults;
+	global  $g_BWWC__config_defaults;
 
-    $bwwc_default_options = $g_BWWC__config_defaults;
+	$bwwc_default_options = $g_BWWC__config_defaults;
 
-    // This will overwrite default options with already existing options but leave new options (in case of upgrading to new version) untouched.
-    $bwwc_settings = BWWC__get_settings ();
-    if (is_array ($bwwc_settings))
-      {
-      foreach ($bwwc_settings as $key=>$value)
-         $bwwc_default_options[$key] = $value;
-      }
+	// This will overwrite default options with already existing options but leave new options (in case of upgrading to new version) untouched.
+	$bwwc_settings = BWWC__get_settings ();
+	if (is_array ($bwwc_settings))
+	{
+	 	foreach ($bwwc_settings as $key=>$value)
+	  	$bwwc_default_options[$key] = $value;
+ 	}
 
-   update_option (BWWC_SETTINGS_NAME, $bwwc_default_options);
+	update_option (BWWC_SETTINGS_NAME, $bwwc_default_options);
 
-   // Re-get new settings.
-   $bwwc_settings = BWWC__get_settings ();
+	// Re-get new settings.
+	$bwwc_settings = BWWC__get_settings ();
 
-   // Create necessaery database tables if not already exists...
-   BWWC__create_database_tables ($bwwc_settings);
+	// Create necessary database tables if not already exists...
+	BWWC__create_database_tables ($bwwc_settings);
+
+	//----------------------------------
+	// Setup cron jobs
+
+	if ($bwwc_settings['enable_soft_cron_job'] && !wp_next_scheduled('BWWC_cron_action'))
+	{
+		$cron_job_schedule_name = strpos($_SERVER['HTTP_HOST'], 'ttt.com')===FALSE ? $bwwc_settings['soft_cron_job_schedule_name'] : 'seconds_30';
+		wp_schedule_event(time(), $cron_job_schedule_name, 'BWWC_cron_action');
+	}
+	//----------------------------------
+
 }
+//---------------------------------------------------------------------------
+// Cron Subfunctions
+function BWWC__add_custom_scheduled_intervals ($schedules)
+{
+	$schedules['seconds_30']     = array('interval'=>30,     'display'=>__('Once every 30 seconds'));     // For testing only.
+	$schedules['minutes_2.5']    = array('interval'=>2.5*60, 'display'=>__('Once every 2.5 minutes'));
+	$schedules['minutes_5']      = array('interval'=>5*60,   'display'=>__('Once every 5 minutes'));
+
+	return $schedules;
+}
+//---------------------------------------------------------------------------
 //===========================================================================
 
 //===========================================================================
@@ -67,7 +92,11 @@ function BWWC_activate()
 function BWWC_deactivate ()
 {
     // Do deactivation cleanup. Do not delete previous settings in case user will reactivate plugin again...
-    // ...
+
+   //----------------------------------
+   // Clear cron jobs
+   wp_clear_scheduled_hook ('BWWC_cron_action');
+   //----------------------------------
 }
 //===========================================================================
 
@@ -75,11 +104,16 @@ function BWWC_deactivate ()
 // uninstalling
 function BWWC_uninstall ()
 {
-    // delete all settings.
-    delete_option(BWWC_SETTINGS_NAME);
+    $bwwc_settings = BWWC__get_settings();
 
-    // delete all tables and data.
-    BWWC__delete_database_tables ();
+    if ($bwwc_settings['delete_db_tables_on_uninstall'])
+    {
+        // delete all settings.
+        delete_option(BWWC_SETTINGS_NAME);
+
+        // delete all DB tables and data.
+        BWWC__delete_database_tables ();
+    }
 }
 //===========================================================================
